+++ src/main/java/com/example/shopping/domain/order/entity/OrderItem.java
@@ -26,6 +26,6 @@ public class OrderItem {
     @JoinColumn(name = "order_id", nullable = false)
     private Order order;
 
-    private Long price;
-    private Long quantity;
+    private Integer price;
+    private Integer quantity;
 }
