+++ src/main/java/com/example/shopping/domain/product/entity/Product.java
@@ -2,7 +2,6 @@ package com.example.shopping.domain.product.entity;
 
 import com.example.shopping.domain.common.entity.TimeStamped;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -19,11 +18,11 @@ public class Product extends TimeStamped {
 
     private String name;
     private String description;
-    private Long price;
+    private Integer price;
     private Long stock;
     private Long viewCount;
 
-    public Product(String name, String description, Long price, Long stock) {
+    public Product(String name, String description, Integer price, Long stock) {
         this.name = name;
         this.description = description;
         this.price = price;
@@ -32,7 +31,7 @@ public class Product extends TimeStamped {
     }
 
     // product 수정
-    public void updateProduct (String name, String description, Long price, Long stock) {
+    public void updateProduct (String name, String description, Integer price, Long stock) {
         this.name = name;
         this.description = description;
         this.price = price;
